requirements.txt
git-python==1.0.3